Moving proxy dumper instantiation from container builder to kernel
